Add files via upload
Cryptocurrencies like Bitcoin, Ethereum and others gained a lot of attention in 2017 when they reached new heights, Whereas Dogecoin gained attention in 2021. Cryptocurrencies are volatile as there is speculation whether they are good investment or not. Given the high variance in prices, can data and statistical analytics methods can be used to provide better investment opportunities to general public?